Added Users and DTOS folder and renamed some folders.
