+++ frazzes-platslageri-be/Application/Application.csproj
@@ -16,4 +16,9 @@
     <ProjectReference Include="..\Domain\Domain.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="DTOS\" />
+    <Folder Include="Users\" />
+  </ItemGroup>
+
 </Project>
