+++ frazzes-platslageri-be/Domain/Models/User.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Identity;
 
-namespace Domain.Model
+namespace Domain.Models
 {
     public class User : IdentityUser<Guid>
     {
