+++ frazzes-platslageri-be/Infrastructure/DependecyInjection/DependecyInjection.cs
@@ -1,5 +1,6 @@
-﻿using Domain.Interface;
-using Infrastructure.Data.Repository;
+﻿using Domain.Interfaces;
+using Domain.Models;
+using Infrastructure.Data.Repositories;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Infrastructure
@@ -8,7 +9,7 @@ namespace Infrastructure
     {
         public static IServiceCollection AddInfrastructure (this IServiceCollection services)
         {
-            services.AddScoped<ICrudRepository, UserRepository>();
+            services.AddScoped<ICrudRepository<User>, UserRepository>();
 
             return services;
         }
